window stuff

+++ source/TF6-MultiFix/YgWindow.c
@@ -6,7 +6,7 @@
 #include "YgWindow.h"
 
 uintptr_t(*_YgSelWnd_Init)(YgSelWnd* window) = (uintptr_t(*)(YgSelWnd*))(0);
-uintptr_t(*_YgSelWnd_Cont)(YgSelWnd* window) = (uintptr_t(*)(YgSelWnd*))(0);
+int(*_YgSelWnd_Cont)(YgSelWnd* window) = (int(*)(YgSelWnd*))(0);
 uintptr_t(*_YgSelWnd_Term)(YgSelWnd* window) = (uintptr_t(*)(YgSelWnd*))(0);
 uintptr_t(*_YgSelWnd_Draw)(uintptr_t ehpacket, YgSelWnd* window) = (uintptr_t(*)(uintptr_t, YgSelWnd*))(0);
 
@@ -41,7 +41,7 @@ uintptr_t YgSelWnd_Init(YgSelWnd* window)
     return _YgSelWnd_Init(window);
 }
 
-uintptr_t YgSelWnd_Cont(YgSelWnd* window)
+int YgSelWnd_Cont(YgSelWnd* window)
 {
     return _YgSelWnd_Cont(window);
 }
@@ -158,7 +158,7 @@ int ygBasicWindow_GetWndState(ygWindowResource* res, ygBasicWindow* window)
 void YgWindow_Init(uintptr_t base_addr)
 {
     _YgSelWnd_Init = (uintptr_t(*)(YgSelWnd*))(0x67A0 + base_addr);
-    _YgSelWnd_Cont = (uintptr_t(*)(YgSelWnd*))(0x7534 + base_addr);
+    _YgSelWnd_Cont = (int(*)(YgSelWnd*))(0x7534 + base_addr);
     _YgSelWnd_Term = (uintptr_t(*)(YgSelWnd*))(0x5CF0 + base_addr);
     _YgSelWnd_Draw = (uintptr_t(*)(uintptr_t, YgSelWnd*))(0x6A84 + base_addr);
 #ifndef YGWINDOW_ESSENTIALS_ONLY
